refs #395, refs #411 fixes found in course of DBus testing * set infobar status correctly when GUI connects after the core has been started * more detailed information in information statusbar * allow to obtain currently connected server in network context * made many network vatlib slots normal member functions * fixed rendered aircraft flag setting in FSX driver

+++ src/blackgui/components/infobarstatuscomponent.h
@@ -55,10 +55,10 @@ namespace BlackGui
 
         private slots:
             //! Simulator connection has been changed
-            void ps_onSimulatorStatusChanged(quint8 status);
+            void ps_onSimulatorStatusChanged(int status);
 
             //! Network connection has been changed
-            void ps_onNetworkConnectionChanged(uint from, uint to);
+            void ps_onNetworkConnectionChanged(int from, int to);
 
             //! Context menu requested
             void ps_customAudioContextMenuRequested(const QPoint &position);
